# docs/swagger.yaml
basePath: /
definitions:
  common.PageResp:
    properties:
      content: {}
      total:
        type: integer
    type: object
  common.jsonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handles.StorageResp:
    properties:
      addition:
        description: Additional information, defined in the corresponding driver
        type: string
      cache_expiration:
        description: cache expire time
        type: integer
      custom_cache_policies:
        type: string
      disable_index:
        type: boolean
      disable_proxy_sign:
        description: Disable sign for DownProxyURL
        type: boolean
      disabled:
        description: if disabled
        type: boolean
      down_proxy_url:
        type: string
      driver:
        description: driver used
        type: string
      enable_sign:
        type: boolean
      extract_folder:
        type: string
      id:
        description: unique key
        type: integer
      modified:
        type: string
      mount_details:
        $ref: '#/definitions/model.StorageDetails'
      mount_path:
        description: must be standardized
        type: string
      order:
        description: use to sort
        type: integer
      order_by:
        type: string
      order_direction:
        type: string
      proxy_range:
        type: boolean
      remark:
        type: string
      status:
        type: string
      web_proxy:
        type: boolean
      webdav_policy:
        type: string
    required:
    - mount_path
    type: object
...
  model.Storage:
    properties:
      addition:
        description: Additional information, defined in the corresponding driver
        type: string
      cache_expiration:
        description: cache expire time
        type: integer
      custom_cache_policies:
        type: string
      disable_index:
        type: boolean
      disable_proxy_sign:
        description: Disable sign for DownProxyURL
        type: boolean
      disabled:
        description: if disabled
        type: boolean
      down_proxy_url:
        type: string
      driver:
        description: driver used
        type: string
      enable_sign:
        type: boolean
      extract_folder:
        type: string
      id:
        description: unique key
        type: integer
      modified:
        type: string
      mount_path:
        description: must be standardized
        type: string
      order:
        description: use to sort
        type: integer
      order_by:
        type: string
      order_direction:
        type: string
      proxy_range:
        type: boolean
      remark:
        type: string
      status:
        type: string
      web_proxy:
        type: boolean
      webdav_policy:
        type: string
    required:
    - mount_path
    type: object
  model.StorageDetails:
    properties:
      totalSpace:
        format: int64
        type: integer
      usedSpace:
        format: int64
        type: integer
    type: object
...
info:
  contact:
    name: OpenList Team
    url: https://github.com/OpenListTeam/
  description: A file list program that supports multiple storage.
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://doc.oplist.org/terms
  title: OpenList
  version: latest
paths:
...
  /api/admin/storage/create:
    post:
      consumes:
      - application/json
      description: Create a new storage
      parameters:
      - description: Storage data
        in: body
        name: storage
        required: true
        schema:
          $ref: '#/definitions/model.Storage'
      produces:
      - application/json
      responses:
        "200":
          description: Success message with new storage ID
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  properties:
                    id:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Storage
      tags:
      - Admin
  /api/admin/storage/delete:
    post:
      description: Delete a storage by ID
      parameters:
      - description: Storage ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Storage
      tags:
      - Admin
  /api/admin/storage/disable:
    post:
      description: Disable a storage by ID
      parameters:
      - description: Storage ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Disable Storage
      tags:
      - Admin
  /api/admin/storage/enable:
    post:
      description: Enable a storage by ID
      parameters:
      - description: Storage ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Enable Storage
      tags:
      - Admin
  /api/admin/storage/get:
    get:
      description: Get storage details by ID
      parameters:
      - description: Storage ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Storage details
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  $ref: '#/definitions/handles.StorageResp'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Storage
      tags:
      - Admin
  /api/admin/storage/list:
    get:
      consumes:
      - application/json
      description: Get a paginated list of storages
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PageResp'
            - properties:
                content:
                  items:
                    $ref: '#/definitions/handles.StorageResp'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: List all storages
      tags:
      - Admin
  /api/admin/storage/load_all:
    post:
      description: Trigger loading of all enabled storages
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Reload all storages
      tags:
      - Admin
  /api/admin/storage/update:
    post:
      consumes:
      - application/json
      description: Update an existing storage
      parameters:
      - description: Storage data
        in: body
        name: storage
        required: true
        schema:
          $ref: '#/definitions/model.Storage'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Storage
      tags:
      - Admin
...
securityDefinitions:
  Authorization:
    description: Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"